# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exchange API",
        "description": "API documentation for ExchangeProject",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/master-settling/monthly/export": {
            "get": {
                "summary": "Export Master Settling Monthly Records to Excel",
                "description": "Download an Excel file containing monthly master settling records.",
                "operationId": "dddad2a35a202f383762f69a3b709909",
                "responses": {
                    "200": {
                        "description": "Excel file downloaded successfully."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/master-settling/weekly/export": {
            "get": {
                "summary": "Export Master Settling Weekly Records to Excel",
                "description": "Download an Excel file containing weekly master settling records.",
                "operationId": "6ead8e9be57ca5117c7fe0a3445698ef",
                "responses": {
                    "200": {
                        "description": "Excel file downloaded successfully."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/master-settling": {
            "get": {
                "summary": "List Master Settling Records for Admin",
                "description": "Retrieve a paginated list of master settling records for admin users.",
                "operationId": "e31f1ee84c416c0b80a61a63ba184c1c",
                "responses": {
                    "200": {
                        "description": "Successful response."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            },
            "post": {
                "summary": "Create a new Master Settling Record",
                "description": "Store a new master settling record in the database.",
                "operationId": "9745d76b568e1e9e3b64f26dffc569bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "white_label",
                                    "credit_reff",
                                    "settling_point",
                                    "price"
                                ],
                                "properties": {
                                    "white_label": {
                                        "type": "string"
                                    },
                                    "credit_reff": {
                                        "type": "string"
                                    },
                                    "settling_point": {
                                        "type": "number"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Master Settling saved successfully."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/assistant/master-settling": {
            "get": {
                "summary": "List Master Settling Records for Assistant",
                "description": "Retrieve a paginated list of master settling records for assistant users.",
                "operationId": "0187d32d0d14de94c18d36c96dc70b3c",
                "responses": {
                    "200": {
                        "description": "Successful response."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/exchange/master-settling": {
            "get": {
                "summary": "List Master Settling Records for Exchange Users",
                "description": "Retrieve a paginated list of master settling records for exchange users.",
                "operationId": "c5174be096ac5a23e67387c75e26ef0a",
                "responses": {
                    "200": {
                        "description": "Successful response."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/master-settling/{id}": {
            "put": {
                "summary": "Update a Master Settling Record",
                "description": "Update an existing master settling record in the database.",
                "operationId": "c2b7964d4842b49c3020e8d39cfa6d1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "white_label",
                                    "credit_reff",
                                    "settling_point",
                                    "price"
                                ],
                                "properties": {
                                    "white_label": {
                                        "type": "string"
                                    },
                                    "credit_reff": {
                                        "type": "string"
                                    },
                                    "settling_point": {
                                        "type": "number"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Master Settling updated successfully."
                    },
                    "404": {
                        "description": "Master Settling not found."
                    }
                }
            },
            "delete": {
                "summary": "Delete a Master Settling Record",
                "description": "Delete a specific master settling record by its ID.",
                "operationId": "3a5eff8a8eb1d192cf77df8a8a651d26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Master Settling deleted successfully."
                    },
                    "404": {
                        "description": "Master Settling not found."
                    }
                }
            }
        },
...
    },
...
}